third_party/blink/web_tests/external/wpt/fetch/private-network-access/iframe.tentative.https.window_include=from-treat-as-public-expected.txt

This is a testharness.js-based test.
[FAIL] treat-as-public-address to local: failed preflight.
  assert_equals: expected "timeout" but got "loaded"
[PASS] treat-as-public-address to local: missing CORS headers.
[PASS] treat-as-public-address to local: missing PNA header.
[FAIL] treat-as-public-address to local: success.
  assert_equals: expected "loaded" but got "timeout"
[PASS] treat-as-public-address to local (same-origin): no preflight required.
[FAIL] treat-as-public-address to private: failed preflight.
  assert_equals: expected "timeout" but got "loaded"
[PASS] treat-as-public-address to private: missing CORS headers.
[PASS] treat-as-public-address to private: missing PNA header.
[FAIL] treat-as-public-address to private: success.
  assert_equals: expected "loaded" but got "timeout"
[PASS] treat-as-public-address to public: no preflight required.
[PASS] treat-as-public-address to local: optional preflight
Harness: the test ran to completion.
